Add support for IDynamicInterfaceCastable-based interfaces in RCW creation
